services/mqtt-kafka-gateway/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
kafka-python
prometheus-client
python-dotenv